refactoring for cleanup

+++ indicators/Hurst/Hurst.Models.cs
@@ -5,6 +5,6 @@ namespace Skender.Stock.Indicators
     [Serializable]
     public class HurstResult : ResultBase
     {
-        public decimal? HurstExponent { get; set; }
+        public double? HurstExponent { get; set; }
     }
 }
